// client/src/app/sharedModels/userDetails.model.ts
export class UserDetails{

    aap_ids: number;
    aap_name: string;
    aap_name_english: string;
    aap_username: string; 
    aap_address: string;
    aap_town: string;
    aap_contact: string;
    aap_profile: string;
    aap_type: string;
    aap_photo: string;
    activation_status: string;
    token: string;

    constructor() {}
}